Mention Rau's work as well.
Truth be said Lam's work is the harder version of software pipelining.
Rau's work pertains to

"Generalized
Vector Computations, (GVC), correspond to loops
which have the following three properties: (I)
the number of iterations to be performed is known
before execution of the loop begins; (2) the body
of the loop is unconditional; (3) there are no
recurrences, i.e., a value computed by one
iteration is not used by a subsequent one."

Condition 3 considerably simplifies the problem.